修改客户模块代码

+++ src/main/java/org/crown/controller/member/MemberRestController.java
@@ -107,7 +107,9 @@ public class MemberRestController extends SuperController {
     })
     @DeleteMapping("/{id}")
     public ApiResponses<Void> delete(@PathVariable("id") Integer id) {
-        memberService.deleteMember(id);
+        if (id != 1) {
+            memberService.deleteMember(id);
+        }
         return success();
     }
